requirements.txt
torch
torchvision
timm
tensorboard
pyyaml
numpy
pillow
